+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/DocEditor.tsx
@@ -117,6 +117,7 @@ export const DocEditor = ({ doc }: DocEditorProps) => {
               initialContent={provider.document.getXmlFragment(
                 'document-store',
               )}
+              docId={doc.id}
             />
           ) : (
             <BlockNoteEditor doc={doc} provider={provider} />